Moved date formatting functions
